+++ OneBlink.SDK/models/Form.cs
@@ -20,7 +20,12 @@ namespace OneBlink.SDK.Model
             long? id = null,
             string postSubmissionAction = "FORMS_LIBRARY",
             bool isAuthenticated = true,
+            List<FormSubmissionEvent> draftEvents = default(List<FormSubmissionEvent>),
             List<FormSubmissionEvent> submissionEvents = default(List<FormSubmissionEvent>),
+            List<FormSubmissionEvent> paymentEvents = default(List<FormSubmissionEvent>),
+            List<FormSubmissionEvent> schedulingEvents = default(List<FormSubmissionEvent>),
+            List<FormApprovalFlowInstanceStep> approvalSteps = default(List<FormApprovalFlowInstanceStep>),
+            List<FormSubmissionEvent> approvalEvents = default(List<FormSubmissionEvent>),
             bool isMultiPage = false,
             string redirectUrl = null,
             bool isInfoPage = false,
@@ -67,7 +72,46 @@ namespace OneBlink.SDK.Model
             {
                 this.submissionEvents = submissionEvents;
             }
-            this.submissionEvents = submissionEvents;
+            if (draftEvents == default(List<FormSubmissionEvent>))
+            {
+                this.draftEvents = new List<FormSubmissionEvent>();
+            }
+            else
+            {
+                this.draftEvents = draftEvents;
+            }
+            if (paymentEvents == default(List<FormSubmissionEvent>))
+            {
+                this.paymentEvents = new List<FormSubmissionEvent>();
+            }
+            else
+            {
+                this.paymentEvents = paymentEvents;
+            }
+            if (schedulingEvents == default(List<FormSubmissionEvent>))
+            {
+                this.schedulingEvents = new List<FormSubmissionEvent>();
+            }
+            else
+            {
+                this.schedulingEvents = schedulingEvents;
+            }
+            if (approvalEvents == default(List<FormSubmissionEvent>))
+            {
+                this.approvalEvents = new List<FormSubmissionEvent>();
+            }
+            else
+            {
+                this.approvalEvents = approvalEvents;
+            }
+            if (approvalSteps == default(List<FormApprovalFlowInstanceStep>))
+            {
+                this.approvalSteps = new List<FormApprovalFlowInstanceStep>();
+            }
+            else
+            {
+                this.approvalSteps = approvalSteps;
+            }
             this.isMultiPage = isMultiPage;
             this.redirectUrl = redirectUrl;
             this.isInfoPage = isInfoPage;
@@ -113,6 +157,26 @@ namespace OneBlink.SDK.Model
             get; set;
         }
         public bool isAuthenticated
+        {
+            get; set;
+        }
+        public List<FormSubmissionEvent> draftEvents
+        {
+            get; set;
+        }
+        public List<FormSubmissionEvent> schedulingEvents
+        {
+            get; set;
+        }
+        public List<FormSubmissionEvent> paymentEvents
+        {
+            get; set;
+        }
+        public List<FormApprovalFlowInstanceStep> approvalSteps
+        {
+            get; set;
+        }
+        public List<FormSubmissionEvent> approvalEvents
         {
             get; set;
         }