+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+### Added
+
+-   More events to forms
+
+### Changed
+
+-   Drafts on submission events
+
 ## 2.6.1 (2021-12-02)
 
 ### Added